src/icebreaker/pyproject.toml
[tool.poetry]
name = "icebreaker"
version = "0.1.0"
description = "Iceberg metrics monitoring platform"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "icebreaker", from = "src"}]

[tool.poetry.dependencies]
python = "^3.8"
streamlit = "^1.32.0"
duckdb = "^0.10.0"
pandas = "^2.2.0"
plotly = "^5.20.0"
pydantic = "^2.6.0"
PyYAML = "^6.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
black = "^24.2.0"
isort = "^5.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api" 